Face interpolator

ExtendableFEM.FaceInterpolator — Method

function FaceInterpolator(
	[kernel,]
	u_args,
	ops_args;
	Tv = Float64,
	Ti = Int,
	kwargs...)

Generates a FaceInterpolator that evaluates discontinuous function operator of its arguments, possibly further postprocessed via a kernel function of interface function kernel(result, input, qpinfo), into a FEVector of H1Pk living on the faces of the grid with the fitting order and number of components.

Keyword arguments:

name: name for operator used in printouts. Default: ''Projector''

only_interior: only interior faces, interpolation of boundary faces will be zero. Default: false

order: interpolation order (default: match order of applied finite element space). Default: ''auto''

parallel_groups: assemble operator in parallel using CellAssemblyGroups. Default: true

params: array of parameters that should be made available in qpinfo argument of kernel function. Default: nothing

resultdim: dimension of result field (default = length of arguments). Default: 0

verbosity: verbosity level. Default: 0

ExtendableFEMBase.evaluate! — Method

function evaluate!(O::FaceInterpolator{Tv, Ti, UT}, sol; kwargs...)

Evaluates the FaceInterpolator using the blocks of sol as arguments and returns the FEVector with the results.